Uploaded sentenceFormatter.py
This program is based on a prompt as part of Ardit Sulce's Python Mega Course. The instructor challenged students to create a program that captures a user's typed input, reformatting the text to have proper capitalization at the beginning of the sentence and the right punctuation. The input should then be collected into and outputted as a paragraph when the user inputs an end command. The only tools I had is what I've learned in the course thus far and Python documentation. This is my own original solution.